Optimize hydration performance 4/5
PHP Reflection API is known to not be the fastest.
This is OK when only a couple of callls are made but here it is used in a tight loop whereas references to values fetched via Reflection are already available or not even used.

Moving the use as late as possible allows to optimize performance a bit.