Optimize hydration performance 3/5
Rework how the hydrator works for better performance by moving most of the column information gathering before the result iteration
so that the gathering is not done X times.
Additionally, this allows to group column info by dql alias which in turns allows to get easily the column list associated with a given dql alias.
This allows to conver the column data lazily, only when needed.
When iterating over a result set representing a high cardinality relation, this avoid using and converting over and over again each column for a given entity.
The only data we systematically need is the identifiers columns in order to determine if the entity is already present in the identity map.

Lowering the impact of hydration for such high cardinality relation lower the thresold where a 2 step hydration process is required.